% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_secrets}
\alias{check_secrets}
\title{Function to check that required environment variables for APPS authentication are set and valid}
\usage{
check_secrets()
}
\description{
Function to check that required environment variables for APPS authentication are set and valid
}
