% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{global_wrapper}
\alias{global_wrapper}
\title{Wrap Global code for Shiny apps}
\usage{
global_wrapper(code)
}
\arguments{
\item{code}{A block of R code}
}
\value{
The result of evaluating the code block in the Global Environment
}
\description{
Allows for global variable assignment and function definitions to be accessible by the server function.
}
