% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_request}
\alias{make_request}
\title{Make an authenticated API request to CSIAPPS}
\usage{
make_request(
  endpoint,
  method = "GET",
  body = NULL,
  query = list(),
  base_url = SITE_URL(),
  token = Sys.getenv("CSIAPPS_ACCESS_TOKEN"),
  timeout = 20L,
  verbose = FALSE,
  paginate = FALSE,
  max_pages = 50
)
}
\arguments{
\item{endpoint}{API endpoint path.}

\item{method}{HTTP method. Defaults to "GET"}

\item{body}{Optional request body for POST/PUT/PATCH requests; should be an R object that can be serialized to JSON}

\item{query}{Optional list of query parameters to include in the request URL}

\item{base_url}{Base URL for the API; defaults to SITE_URL()}

\item{token}{Authentication token. Will attempt to read from CSIAPPS_ACCESS_TOKEN environment variable if not provided explicitly.}

\item{timeout}{Request timeout in seconds; defaults to 20}

\item{verbose}{If TRUE, prints request and response details to the console for debugging purposes}

\item{paginate}{If TRUE, will attempt to paginate through results using "next" links in the API response. Defaults to FALSE.}

\item{max_pages}{Maximum number of pages to fetch when paginate = TRUE; defaults to 50 to prevent infinite loops}
}
\value{
List of parsed API responses
}
\description{
Make an authenticated API request to CSIAPPS
}
