% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse_client.R
\name{new_warehouse_client}
\alias{new_warehouse_client}
\title{Create a Warehouse API client}
\usage{
new_warehouse_client(config = WarehouseAPIConfig())
}
\arguments{
\item{config}{Configuration list for API endpoints and settings. Defaults to \code{WarehouseAPIConfig()}.}
}
\value{
A list of functions for interacting with the Warehouse API
}
\description{
Create a Warehouse API client
}
