% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkce_state_encode}
\alias{pkce_state_encode}
\title{Encode a PKCE code verifier into a state string for the auth request}
\usage{
pkce_state_encode(verifier)
}
\arguments{
\item{verifier}{a PKCE code verifier string}
}
\description{
Encode a PKCE code verifier into a state string for the auth request
}
