% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{server_wrapper}
\alias{server_wrapper}
\title{Wrapper server function for Shiny apps}
\usage{
server_wrapper(app_specific_logic)
}
\arguments{
\item{app_specific_logic}{Existing server logic of shiny web application}
}
\value{
A Shiny server function that wraps the provided app-specific logic with authentication handling and user info retrieval.
}
\description{
Handles OAuth2 PKCE authentication flow with CSIAPPS, managing user tokens and info, and providing a consistent authentication status UI.
}
