% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse_client.R
\name{WarehouseAPIConfig}
\alias{WarehouseAPIConfig}
\title{Default configuration for Warehouse API client}
\usage{
WarehouseAPIConfig(
  base_url = SITE_URL(),
  ingest_path = "/api/warehouse/ingestion/primary/",
  datasource_detail_pattern = "/api/warehouse/data-sources/\%s/",
  records_list_path = "/api/warehouse/data-records/",
  timeout_s = 20L
)
}
\arguments{
\item{base_url}{the base URL for the Warehouse API}

\item{ingest_path}{the path for the ingestion endpoint}

\item{datasource_detail_pattern}{a pattern for datasource detail endpoint}

\item{records_list_path}{the path for the data records endpoint}

\item{timeout_s}{timeout for API requests in seconds}
}
\value{
A list containing the API configuration
}
\description{
Default configuration for Warehouse API client
}
