% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ui_wrapper}
\alias{ui_wrapper}
\title{Wrapper UI for all Shiny apps, providing a consistent navbar and footer, and handling authentication status display.}
\usage{
ui_wrapper(...)
}
\arguments{
\item{...}{Additional UI elements to include in the main content area}
}
\value{
A Shiny UI object with a navbar, footer, and main content area
}
\description{
Wrapper UI for all Shiny apps, providing a consistent navbar and footer, and handling authentication status display.
}
